Secure_log: add 'send_eof' parameter
Rsyslog reads the files and hangs up until it reads a predetermined
number of lines. This option makes secure_log send a EOF when the
buffer is empty, allowing rsyslog to flush its cache.
This option have security consequences: if someone else is reading
the socket at the same time, some data may not be read by rsyslog