Security_log: security fix, potential buffer overflow at end of buffer
The reader might be positioned at the end of the buffer before
this recored is 'cleaned out' by a new record (if there is not enough
room remaining). However, when reading, the reader does not check if
the record was 'cleaned out' ("len" set to 0). This commit add such a check:
If the len of the current record is 0, go back to the beginning
of the buffer.