Netlog: change probes early loading behavior
Before this commits, probes were immediately loaded as soon as a
kernel parameter was set. This would even be the case before the
module was properly loaded (or even started module_init).

This can results in invalid memory access if the module_init is
not called and the module unloaded (without going through
module_exit). This is not supposed to be possible, but the 'dresden'
hack does result in such a stupid case...

The new behavior delays the probe loading until module_init