secure_log: fix duplicate reading when wrapping around
In secure_log_read, if log_curr_idx points to the end of the buffer and
indicates a wrapping, we directly wrap and read the record at the
beginning of the buffer. However, we don't reset the log_curr_idx.
After the message is read, 'next_record(log_curr_idx) will reset
log_curr_idx, pointing it to the very same message as the one we just
read. On the next run, we will read it again.