Execlog: Fix out of bound memory access on fail path
When execlog fails to allocate enough memory to copy the argv of the user,
an error message is used instead. Unfortunately, on this path, the argv_size
is not updated with the new length and still contains the length that
resulted in the memory allocation.

This was introduced by c4497cbbbb35f7b5d59b8a4fd18f0cc31da60a4c:
"Execlog: Re-evaluate argv_size and store value earlier"