Execlog: fix an off-by-one overflow when truncating argv
When argv is truncated (either if input is bigger than the deined max, 39936)
or if it was resized dynamically between the two calls), the buffer gets fully
written by the user input, leaving no place for the last bit (\0), written
afterwards. Because the buffer is at least of size 2, when it's full we can
safely go back by one, to overwrite the last bit with '\0'.